+++ avo-icons.gemspec
@@ -1,13 +1,13 @@
-require_relative "lib/avo/heroicons/version"
+require_relative "lib/avo/icons/version"
 
 Gem::Specification.new do |spec|
-  spec.name        = "avo-heroicons"
-  spec.version     = Avo::Heroicons::VERSION
+  spec.name        = "avo-icons"
+  spec.version     = Avo::Icons::VERSION
   spec.authors     = [ "Paul Bob", "Adrian Marin" ]
   spec.email       = [ "avo@avohq.io" ]
   spec.homepage    = "https://avohq.io"
-  spec.summary     = "A lightweight Avo dependency for serving Heroicons SVGs effortlessly."
-  spec.description = "A lightweight Avo dependency for serving Heroicons SVGs effortlessly."
+  spec.summary     = "A lightweight Avo dependency for serving icon SVGs effortlessly."
+  spec.description = "A lightweight Avo dependency for serving icon SVGs effortlessly."
 
   spec.metadata["homepage_uri"] = spec.homepage
