+++ lib/avo/icons/version.rb
@@ -1,5 +1,5 @@
 module Avo
-  module Heroicons
+  module Icons
     VERSION = "0.1.2"
   end
 end
