+++ lib/tasks/avo/icons_tasks.rake
