--- test/avo/heroicons_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class Avo::HeroiconsTest < ActiveSupport::TestCase
-  test "it has a version number" do
-    assert Avo::Heroicons::VERSION
-  end
-end
