+++ test/avo/icons_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class Avo::IconsTest < ActiveSupport::TestCase
+  test "it has a version number" do
+    assert Avo::Icons::VERSION
+  end
+end
